flask/requirements.txt
Flask
Werkzeug
cairosvg
svg_to_gcode
lxml
Pillow
gunicorn
flask-cors
watchdog
cssutils